requirements.txt
dj-database-url==0.5.0
Django==2.2.15
django-ckeditor==5.9.0
django-jazzmin==2.2.7
django-js-asset==1.2.2
django-stdimage==5.1.1
django-userforeignkey==0.3.0
mysqlclient==2.0.1
Pillow==7.2.0
progressbar2==3.52.1
PyMySQL==0.10.1
python-decouple==3.3
python-utils==2.4.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
